+++ internal/domain/author/usecase/usecase_mock.go
@@ -11,9 +11,9 @@ import (
 // AuthorMock is a mock implementation of Author.
 type AuthorMock struct {
 	CreateFunc func(ctx context.Context, a *author.CreateRequest) (*author.Schema, error)
-	DeleteFunc func(ctx context.Context, authorID uint) error
+	DeleteFunc func(ctx context.Context, authorID uint64) error
 	ListFunc   func(ctx context.Context, f *author.Filter) ([]*author.Schema, int, error)
-	ReadFunc   func(ctx context.Context, authorID uint) (*author.Schema, error)
+	ReadFunc   func(ctx context.Context, authorID uint64) (*author.Schema, error)
 	UpdateFunc func(ctx context.Context, authorMiripParam *author.UpdateRequest) (*author.Schema, error)
 }
 
@@ -21,7 +21,7 @@ func (m *AuthorMock) Create(ctx context.Context, a *author.CreateRequest) (*auth
 	return m.CreateFunc(ctx, a)
 }
 
-func (m *AuthorMock) Delete(ctx context.Context, authorID uint) error {
+func (m *AuthorMock) Delete(ctx context.Context, authorID uint64) error {
 	return m.DeleteFunc(ctx, authorID)
 }
 
@@ -29,7 +29,7 @@ func (m *AuthorMock) List(ctx context.Context, f *author.Filter) ([]*author.Sche
 	return m.ListFunc(ctx, f)
 }
 
-func (m *AuthorMock) Read(ctx context.Context, authorID uint) (*author.Schema, error) {
+func (m *AuthorMock) Read(ctx context.Context, authorID uint64) (*author.Schema, error) {
 	return m.ReadFunc(ctx, authorID)
 }
